requirements.txt
python-xdg
py-flags
python-urllib3
python-blist
python-orange
Orange3-Associate
